trainer/requirements.txt
torch==2.3.0+cu121     --index-url https://download.pytorch.org/whl/cu121 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.2
pandas-ta==0.3.14b0
tqdm==4.66.4
pytz==2024.1
scikit-learn==1.5.0
